requirements.txt
streamlit
wikipedia-api
google-genai
python-dotenv
pyvis